+++ cmd/preparerelease/internal/key.go
@@ -1,0 +1,6 @@
+package internal
+
+const (
+	FileChangelogMd = "CHANGELOG.md"
+	FileProjectGo   = "pkg/project/project.go"
+)