+++ cmd/preparerelease/internal/modifier_test.go
@@ -1,0 +1,6 @@
+package internal
+
+import "testing"
+
+func Test_modifier_addReleaseToChangelogMd(t *testing.T) {
+}