+++ CHANGELOG.md
@@ -8,6 +8,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2020-06-08
+
+### Added
+
+- Add `prepare-release` command (#442).
+
+## [1.1.3] 2020-06-05
+
+### Changed
+
+-  Update giantswarm/app to 0.2.2 and use 0.0.0 as version for app CRs.
+
+## [1.1.2] 2020-06-03
+
+### Changed
+
+-  Revert giantswarm/app to 0.2.1 and use 1.0.0 as version for app CRs.
+
+## [1.1.1] 2020-06-02
+
+### Changed
+
+-  Update giantswarm/app to 0.2.2 and use 0.0.0 as version for app CRs.
+
 ## [1.1.0] 2020-05-28
 
 ### Changed
@@ -58,6 +82,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add changelog.
 - Add SemVer versioning.
 
-[Unreleased]: https://github.com/giantswarm/architect/compare/v1.1.0...HEAD
-[1.1.0]: https://github.com/giantswarm/architect/releases/tag/v1.1.0
+[Unreleased]: https://github.com/giantswarm/architect/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/giantswarm/architect/compare/v1.1.2...v1.2.0
+[1.1.2]: https://github.com/giantswarm/architect/compare/v1.1.1...v1.1.2
+[1.1.1]: https://github.com/giantswarm/architect/compare/v1.1.0...v1.1.1
+[1.1.0]: https://github.com/giantswarm/architect/compare/v1.0.0...v1.1.0
+
 [1.0.0]: https://github.com/giantswarm/architect/releases/tag/v1.0.0